api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=2.2.2,<3.1.0",
    "werkzeug>=2.2.0,<3.1.0",
    "gunicorn>=21.0.0",
    "flask-sqlalchemy>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "argon2-cffi>=23.1.0",
    "pyjwt>=2.8.0",
    "flask-cors>=4.0.0",
    "flask-limiter>=3.5.0",
    "flask-migrate>=4.0.0",
    "flask-socketio>=5.3.0",
    "eventlet>=0.33.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=25.0.0"]

[tool.black]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
include = ["hearts*"]
exclude = ["migrations*"]
